Extract AbstractStoreOptions with onLoaded member
