+++ packages/devextreme/js/data/odata/store.d.ts
@@ -1,6 +1,5 @@
 import { DxPromise } from '../../core/utils/deferred';
 import Store, { Options as StoreOptions } from '../abstract_store';
-import { LoadOptions } from '../index';
 import { Query } from '../query';
 import { ODataRequestOptions } from './context';
 
@@ -11,9 +10,9 @@ export type Options<
 > = ODataStoreOptions<TItem, TKey>;
 
 /**
+ * @docid
  * @namespace DevExpress.data
  * @deprecated Use Options instead
- * @docid
  */
 export interface ODataStoreOptions<
     TItem = any,
@@ -66,13 +65,6 @@ export interface ODataStoreOptions<
     keyType?: 'String' | 'Int32' | 'Int64' | 'Guid' | 'Boolean' | 'Single' | 'Decimal' | any;
     /**
      * @docid
-     * @type_function_param1 loadOptions:LoadOptions
-     * @action
-     * @public
-     */
-    onLoading?: ((loadOptions: LoadOptions<TItem>) => void);
-    /**
-     * @docid
      * @public
      */
     url?: string;
@@ -90,6 +82,7 @@ export interface ODataStoreOptions<
      */
     withCredentials?: boolean;
 }
+
 /**
  * @docid
  * @inherits Store
