+++ packages/devextreme/js/data/store.d.ts
@@ -1,0 +1,209 @@
+import {
+  DxPromise,
+  DxExtendedPromise,
+} from '../core/utils/deferred';
+import { DeepPartial } from '../core/index';
+import {
+  FilterDescriptor,
+  GroupDescriptor,
+  LoadOptions,
+} from './index';
+
+export type Options<
+    TItem = any,
+    TKey = any,
+> = StoreOptions<TItem, TKey>;
+
+/**
+ * @docid
+ * @namespace DevExpress.data
+ * @deprecated Use Options instead
+ * @hidden
+ */
+export interface StoreOptions<
+    TItem = any,
+    TKey = any,
+> {
+    /**
+     * @docid
+     * @public
+     */
+    errorHandler?: Function;
+    /**
+     * @docid
+     * @public
+     */
+    key?: string | Array<string>;
+    /**
+     * @docid
+     * @type_function_param1 values:object
+     * @type_function_param2 key:object|string|number
+     * @action
+     * @public
+     */
+    onInserted?: ((values: TItem, key: TKey) => void);
+    /**
+     * @docid
+     * @type_function_param1 values:object
+     * @action
+     * @public
+     */
+    onInserting?: ((values: TItem) => void);
+    /**
+     * @docid
+     * @type_function_param1 loadOptions:LoadOptions
+     * @action
+     * @public
+     */
+    onLoading?: ((loadOptions: LoadOptions<TItem>) => void);
+    /**
+     * @docid
+     * @action
+     * @public
+     */
+    onModified?: Function;
+    /**
+     * @docid
+     * @action
+     * @public
+     */
+    onModifying?: Function;
+    /**
+     * @docid
+     * @action
+     * @public
+     */
+    onPush?: ((changes: Array<TItem>) => void);
+    /**
+     * @docid
+     * @type_function_param1 key:object|string|number
+     * @action
+     * @public
+     */
+    onRemoved?: ((key: TKey) => void);
+    /**
+     * @docid
+     * @type_function_param1 key:object|string|number
+     * @action
+     * @public
+     */
+    onRemoving?: ((key: TKey) => void);
+    /**
+     * @docid
+     * @type_function_param1 key:object|string|number
+     * @type_function_param2 values:object
+     * @action
+     * @public
+     */
+    onUpdated?: ((key: TKey, values: TItem) => void);
+    /**
+     * @docid
+     * @type_function_param1 key:object|string|number
+     * @type_function_param2 values:object
+     * @action
+     * @public
+     */
+    onUpdating?: ((key: TKey, values: TItem) => void);
+}
+
+type EventName = 'loaded' | 'loading' | 'inserted' | 'inserting' | 'updated' | 'updating' | 'push' | 'removed' | 'removing' | 'modified' | 'modifying';
+
+/**
+ * @docid Store
+ * @namespace DevExpress.data
+ * @hidden
+ * @options StoreOptions
+ */
+export class Store<
+    TItem = any,
+    TKey = any,
+> {
+    constructor(options?: Options<TItem, TKey>);
+    /**
+     * @docid
+     * @publicName insert(values)
+     * @param1 values:object
+     * @return Promise<any>
+     * @public
+     */
+    insert(values: TItem): DxExtendedPromise<TItem>;
+    /**
+     * @docid
+     * @publicName key()
+     * @public
+     */
+    key(): string | Array<string>;
+    /**
+     * @docid
+     * @publicName keyOf(obj)
+     * @param1 obj:object
+     * @return any|string|number
+     * @public
+     */
+    keyOf(obj: TItem): TKey;
+    /**
+     * @docid
+     * @publicName off(eventName)
+     * @param1 eventName:string
+     * @return this
+     * @public
+     */
+    off(eventName: EventName): this;
+    /**
+     * @docid
+     * @publicName off(eventName, eventHandler)
+     * @param1 eventName:string
+     * @return this
+     * @public
+     */
+    off(eventName: EventName, eventHandler: Function): this;
+    /**
+     * @docid
+     * @publicName on(eventName, eventHandler)
+     * @param1 eventName:string
+     * @return this
+     * @public
+     */
+    on(eventName: EventName, eventHandler: Function): this;
+    /**
+     * @docid
+     * @publicName on(events)
+     * @param1 events:object
+     * @return this
+     * @public
+     */
+    on(events: { [key in EventName]?: Function }): this;
+    /**
+     * @docid
+     * @publicName push(changes)
+     * @param1 changes:Array<any>
+     * @public
+     */
+    push(changes: Array<{ type: 'insert' | 'update' | 'remove'; data?: DeepPartial<TItem>; key?: TKey; index?: number }>): void;
+    /**
+     * @docid
+     * @publicName remove(key)
+     * @param1 key:object|string|number
+     * @return Promise<void>
+     * @public
+     */
+    remove(key: TKey): DxPromise<void>;
+    /**
+     * @docid
+     * @publicName totalCount(options)
+     * @param1_field filter:object
+     * @param1_field group:object
+     * @return Promise<number>
+     * @public
+     */
+    totalCount(obj: { filter?: FilterDescriptor | Array<FilterDescriptor>; group?: GroupDescriptor<TItem> | Array<GroupDescriptor<TItem>> }): DxPromise<number>;
+    /**
+     * @docid
+     * @publicName update(key, values)
+     * @param1 key:object|string|number
+     * @param2 values:object
+     * @return Promise<any>
+     * @public
+     */
+    update(key: TKey, values: DeepPartial<TItem>): DxExtendedPromise<TItem>;
+}
